Move general impl
